+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.0
+
+Report an error when a file that contains components that can't be fast-refreshed because:
+
+- There are no export (entrypoint)
+- Exports are not components
+
 ## 0.2.1
 
 Doc only: Update limitations section README.md